fix: job finish queue events race condition
This fixes a bug where waiting for a job to complete can timeout if the job finished faster than the listeners can be registered.

Before this change, `job.waitUntilFinished` did a smart check at the start to see if the job had already been completed. This makes sense -- the QueueEvents object is never going to send another `completed` event for a job that has already been completed, so we need to make sure that we're not listening for an event that never comes.

The problem is that there's a finite amount of time between when we make this first check and when we start listening to the queue events. There is a window after we make the check and before we are listening for the completed event. For very quick to process jobs, they can emit their completed event in this window, so the listeners never find it. The `waitUntilFinished` will fail with a timeout, but the job did actually complete.

This switches the order in which we do things to eliminate this window: we add the completed event listeners first, and then poll once. That way, the jobs that finish very quickly will either hit the event listener if it's already been added, or be noticed by the check at the end. Jobs that have already been completed will be noticed by the check at the end. `Promise`es can only resolve once, so even if both checks fire the function will only return once.

The downside of this approach is that for jobs that have already been completed before the function is even called, we add some extra listeners to the QueueEvents that didn't strictly need to be added. Since these are just in memory doodads, I think this is ok. An alternative would be to run `Scripts.isFinished` at the start and end of the function, but I think that's worse and causes more redis load, whereas extra EventEmitter listeners don't.

One other caveat is that the `QueueEvents` object passed into `waitUntilFinished` needs to be fully booted before you can trust that no events would be missed. Make sure to await `QueueEvents.waitForReady` before using it with `waitUntilFinished`.