fix: ensure connection reconnects when pausing
Currently, if we pause a worker without explicitly requesting to no wait on
active jobs, we potentially disconnect (assuming we're currently already waiting
for a job). The issue is that `this.waiting` can change to `false` before we try
to reconnect, preventing us from reconnecting at all. Once we resume, we're no
longer paused or closing, but the connections is closed, causing the `run` loop
to throw `Client is disconnected`.

We fix this by _only_ checking the `reconnect` argument, but setting it to false
if we didn't disconnect.