feat(child-pool): Helpful Error on Interpolation
If you have an error in your sandboxed code that a causes NodeJS to crash (ie. doing a require to a file that does not exist) you get an un-helpful error `Error initializing child: Internal Exception Handler Run-Time Failure and signal null`.

This change propagates the actual error up as a job failure.