feat(worker.ts): adds new sequential job fetch mode
Adding new 'sequential' job fetch mode that can be used in scenarios where
high worker concurrency mode is wanted but we want to avoid putting too
much pressure on Redis.
In this mode, worker does not issue `concurrency` fetch jobs queries to
Redis in parallel, and instead loads them sequentially. This does not mean
that jobs are processed sequentially!

re #2157